Addition of missing arduino debugs #210

+++ cpp_utils/BLEAdvertisedDevice.cpp
@@ -209,7 +209,7 @@ void BLEAdvertisedDevice::parseAdvertisement(uint8_t* payload) {
 		payload++; // Skip to type
 		sizeConsumed += 1 + length; // increase the size consumed.
 
-		if (length != 0) { // A length of 0 indicate that we have reached the end.
+		if (length != 0) { // A length of 0 indicates that we have reached the end.
 			ad_type = *payload;
 			payload++;
 			length--;
@@ -253,7 +253,7 @@ void BLEAdvertisedDevice::parseAdvertisement(uint8_t* payload) {
 				case ESP_BLE_AD_TYPE_32SRV_CMPL:
 				case ESP_BLE_AD_TYPE_32SRV_PART: { // Adv Data Type: 0x04
 					for (int var = 0; var < length/4; ++var) {
-						setServiceUUID(BLEUUID(*reinterpret_cast<uint16_t*>(payload+var*4)));
+						setServiceUUID(BLEUUID(*reinterpret_cast<uint32_t*>(payload+var*4)));
 					}
 					break;
 				} // ESP_BLE_AD_TYPE_32SRV_PART
@@ -362,13 +362,15 @@ void BLEAdvertisedDevice::setScan(BLEScan* pScan) {
 	m_pScan = pScan;
 } // setScan
 
+
 /**
  * @brief Set the Service UUID for this device.
  * @param [in] serviceUUID The discovered serviceUUID
  */
 void BLEAdvertisedDevice::setServiceUUID(const char* serviceUUID) {
 	return setServiceUUID(BLEUUID(serviceUUID));
-} // setRSSI
+} // setServiceUUID
+
 
 /**
  * @brief Set the Service UUID for this device.
@@ -378,7 +380,7 @@ void BLEAdvertisedDevice::setServiceUUID(BLEUUID serviceUUID) {
 	m_serviceUUIDs.push_back(serviceUUID);
 	m_haveServiceUUID = true;
 	ESP_LOGD(LOG_TAG, "- addServiceUUID(): serviceUUID: %s", serviceUUID.toString().c_str());
-} // setRSSI
+} // setServiceUUID
 
 
 /**
